[#4] [Chore] Move interfaces and concrete implementations to separate files
