+++ lib/src/main/java/co/nimblehq/alpine/lib/mrz/MrzProcessor.kt
@@ -1,0 +1,10 @@
+package co.nimblehq.alpine.lib.mrz
+
+interface MrzProcessor {
+    fun processImageFile(filePath: String, mrzProcessorResultListener: MrzProcessorResultListener)
+
+    companion object {
+        @JvmStatic
+        fun newInstance(): MrzProcessor = MrzProcessorImpl()
+    }
+}
