+++ lib/src/main/java/co/nimblehq/alpine/lib/nfc/NfcReader.kt
@@ -1,0 +1,19 @@
+package co.nimblehq.alpine.lib.nfc
+
+import android.content.Context
+import android.nfc.Tag
+import androidx.annotation.WorkerThread
+import co.nimblehq.alpine.lib.model.MrzInfo
+import co.nimblehq.alpine.lib.model.PassportInfo
+
+interface NfcReader {
+    @WorkerThread
+    fun readNfc(tag: Tag, mrzInfo: MrzInfo, timeout: Int = ISO_DEP_TIMEOUT_IN_MILLIS): PassportInfo?
+
+    companion object {
+        private const val ISO_DEP_TIMEOUT_IN_MILLIS = 60000
+
+        @JvmStatic
+        fun newInstance(context: Context): NfcReader = NfcReaderImpl(context)
+    }
+}
